good run through

+++ reddit-comment-image-getter.php
@@ -8,6 +8,11 @@ By Stan James http://wanderingstan.com
 
 include 'jpg_url_from_service_url.php';
 
+function clean_filename($filename) {
+	$replace="_";
+	$pattern="/([[:alnum:]_\.-]*)/";
+	return str_replace(str_split(preg_replace($pattern,$replace,$fname)),$replace,$filename);
+}
 
 function download_comment_images_from_post($post_url) {
 
@@ -20,8 +25,7 @@ function download_comment_images_from_post($post_url) {
 
 	echo $post_title . "\n";
 	// try to get word of the day
-	$pattern = '/WotD "([^"]*)"/';
-	$post_title_re = '/WotD *"([^"]*)([ "]*)?([0-9]+\/[0-9]+)?/';
+	$post_title_re = '/WotD[^"]*"([^"]*)([ "]*)?([0-9]+\/[0-9]+)?/';
 
 	preg_match($post_title_re, $post_title, $matches, PREG_OFFSET_CAPTURE);
 	$wotd = $matches[1][0];
@@ -34,7 +38,7 @@ function download_comment_images_from_post($post_url) {
 	}
 
 	// create director for posts
-	$post_directory_name = $post_created_date . "-" . $wotd;
+	$post_directory_name = $post_created_date . "-" . clean_filename($wotd);
 	if (!file_exists($post_directory_name)) {
 	    mkdir($post_directory_name, 0777, true);
 	}
@@ -58,21 +62,26 @@ function download_comment_images_from_post($post_url) {
 				error_log("Err: More than one image found in comment: " . $comment_body);
 			}
 			echo $image_filename . "\n";
-			$img_html_url = $match[0][0]; // URL of the "image html page" on hosted service. NOT the actual jpg. 
-			$img_url = jpg_url_from_service_url($img_html_url);
-			if (!$img_url) {
-				error_log  ("Err: Could not extract image from: " . $img_html_url);
-			}
-			else {
-				// Download our image and savev to local file
-				file_put_contents($image_filename, file_get_contents($img_url));						
-			}
 
-			sleep(2);
+			if (!file_exists($image_filename)) {
+				$img_html_url = $match[0][0]; // URL of the "image html page" on hosted service. NOT the actual jpg. 
+				$img_url = jpg_url_from_service_url($img_html_url);
+				if (!$img_url) {
+					error_log  ("Err: Could not extract image from: " . $img_html_url);
+				}
+				else {
+					// Download our image and savev to local file
+					file_put_contents($image_filename, file_get_contents($img_url));						
+				}
+				// be polite and sleep between requests
+				sleep(2);
+			}
 		}
 	}
 
-	shell_exec ( 'montage '.$post_directory_name.'/*.jpg -geometry 150x150+2+2 -tile 2x '.$montage_directory_name.'/'.$post_directory_name.'.jpg' );
+	// shell_exec ( 'montage '.$post_directory_name.'/*.jpg -geometry 150x150+2+2 -tile 2x -background black '.$montage_directory_name.'/'.$post_directory_name.'.jpg' );
+
+	shell_exec ( 'montage '.$post_directory_name.'/*.jpg -geometry 300x300+2+2 -tile 2x -background black '.$montage_directory_name.'/'.$post_directory_name.'.jpg' );
 
 }
